Separating mocks from tests and services
